pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "incae-llm-seminar"
version = "0.1.0"
description = "Runnable materials for INCAE LLM Seminar sessions on text classification."
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "pandas>=2.1",
  "pyyaml>=6.0.1",
  "torch>=2.2",
  "transformers>=4.39",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.1",
]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
